Frame Equalizer Update
Made the hardcoded private variables into raw pointers that allocate
space based on user input and changed the interleaver pattern so that it
should calculate the pattern on the fly. It is based on stripping a row
so of subcarriers so it works with multiples of 3, no protection against
numbers not a multiple of 3.